Use Nextjs image component for lazy loading images
We previously disabled this because our "optimize" images quote ran out.
This is some Vercel-image-CDN magic that will optimize images as clients
request them.

However, I just realized we can add an `unoptimized` prop to the Image
component, that supposedly will skip this step.

All our images are semi-optmized anyways, so it shouldn't matter. In the
future, we might move non-album images to Sanity as well, which will
give us image CDN without the extra cost.